♻️ refactor: extract GLSL shaders into separate files (#9)
- .glsl files for syntax highlighting and GLSL tooling
- shaders/index.ts exports string constants for bundling
- glsl.d.ts module declaration for .glsl imports
- webgl-renderer.ts imports from shaders/ instead of inline strings
