+++ packages/core/src/renderers/shaders/glsl.d.ts
@@ -1,0 +1,5 @@
+/** Allow importing .glsl files as raw strings */
+declare module '*.glsl' {
+  const source: string;
+  export default source;
+}
